CI: update doc preview config (#974)
* CI: only doc preview pull requests

When merging to master, the action fails with
"[HttpError] Not found", showing a URL to
the pull requests for the repository.

The documentation at

https://github.com/readthedocs/actions/tree/v1/preview

does not suggest running on merges,
so only run on pull requests.

* CI: use pull_request_target for docpreview

The documentation at

https://github.com/readthedocs/actions/tree/v1/preview

uses this. According to

https://stackoverflow.com/questions/74957218/what-is-the-difference-between-pull-request-and-pull-request-target-event-in-git

there is something with permissions of tokens
that differ between pull_request and pull_request_target,
so switch to what the action documentation
suggests.