wcs?_utils: fixes + some type signatures
The bands member is a list, so convert
the set to a list.

The width/height are integers, so convert
to int directly when doing division instead
of first writing a float to the members.

MyPy complains about missing things when
formatting the exception message, but it
seems to understand the situation when
changing that access for the attribute
directly, so do that.

The remaining issue is that "output" can be
None, but ignore that for now and add a FIXME.